refactor: consolidate activity services and implement repository pattern - Introduced ActivityRepository to handle data access for activities. - Unified activities service logic into a single ActivityService class. - Updated Activity router and CloseCycle service to use the new unified service. - Removed obsolete individual service files from `app/services/activities/`. - Aligned `activities` package structure with the `users` package refactoring.
